Allow changing of data size in ByteContainer.setBytes(byte[], short, short).
The motivation for this patch is https://github.com/ANSSI-FR/SmartPGP/issues/40.

However, this workaround avoids looking at the underlying problem:
ECKeyImpl(byte, short) should not initialize the DomainParameters (according to its own docs), but does.
This leads to problems with NIST P-521 because the default parameter for b has a leading zero bit that gets stripped, leading to a change in data size, when SmartPGP later sets b explicitly with the leading zero bit intact.